+++ src/container/HashContainer/HashMap.ts
@@ -1,10 +1,10 @@
-import { initContainer } from '@/container/ContainerBase/index';
+import HashContainer from './Base/index';
 import Vector from '../SequentialContainer/Vector';
 import OrderedMap from '../TreeContainer/OrderedMap';
-import HashContainerBase from './Base/index';
+import { initContainer } from '@/container/ContainerBase/index';
 
-class HashMap<K, V> extends HashContainerBase<K> {
-  private hashTable: (Vector<[K, V]> | OrderedMap<K, V>)[] = [];
+class HashMap<K, V> extends HashContainer<K> {
+  protected hashTable: (Vector<[K, V]> | OrderedMap<K, V>)[] = [];
   constructor(
     container: initContainer<[K, V]> = [],
     initBucketNum? :number,
@@ -12,15 +12,19 @@ class HashMap<K, V> extends HashContainerBase<K> {
     super(initBucketNum, hashFunc);
     container.forEach(element => this.setElement(element[0], element[1]));
   }
-  private reAllocate() {
-    if (this.bucketNum >= HashContainerBase.maxBucketNum) return;
+  protected reAllocate() {
+    if (this.bucketNum >= HashContainer.maxBucketNum) return;
     const newHashTable: (Vector<[K, V]> | OrderedMap<K, V>)[] = [];
     const originalBucketNum = this.bucketNum;
     this.bucketNum <<= 1;
-    for (let index = 0; index < originalBucketNum; ++index) {
+    const keys = Object.keys(this.hashTable);
+    const keyNums = keys.length;
+    for (let i = 0; i < keyNums; ++i) {
+      const index = parseInt(keys[i]);
       const container = this.hashTable[index];
-      if (!container || container.empty()) continue;
-      if (container.size() === 1) {
+      const size = container.size();
+      if (size === 0) continue;
+      if (size === 1) {
         const element = container.front() as [K, V];
         newHashTable[
           this.hashFunc(element[0]) & (this.bucketNum - 1)
@@ -29,30 +33,30 @@ class HashMap<K, V> extends HashContainerBase<K> {
       }
       const lowList: [K, V][] = [];
       const highList: [K, V][] = [];
-      container.forEach(pair => {
-        const hashCode = this.hashFunc(pair[0]);
+      container.forEach(element => {
+        const hashCode = this.hashFunc(element[0]);
         if ((hashCode & originalBucketNum) === 0) {
-          lowList.push(pair);
-        } else highList.push(pair);
+          lowList.push(element);
+        } else highList.push(element);
       });
       if (container instanceof OrderedMap) {
-        if (lowList.length > HashMap.untreeifyThreshold) {
+        if (lowList.length > HashContainer.untreeifyThreshold) {
           newHashTable[index] = new OrderedMap(lowList);
         } else if (lowList.length) {
           newHashTable[index] = new Vector(lowList, false);
         }
-        if (highList.length > HashMap.untreeifyThreshold) {
+        if (highList.length > HashContainer.untreeifyThreshold) {
           newHashTable[index + originalBucketNum] = new OrderedMap(highList);
         } else if (highList.length) {
           newHashTable[index + originalBucketNum] = new Vector(highList, false);
         }
       } else {
-        if (lowList.length >= HashMap.treeifyThreshold) {
+        if (lowList.length >= HashContainer.treeifyThreshold) {
           newHashTable[index] = new OrderedMap(lowList);
         } else if (lowList.length) {
           newHashTable[index] = new Vector(lowList, false);
         }
-        if (highList.length >= HashMap.treeifyThreshold) {
+        if (highList.length >= HashContainer.treeifyThreshold) {
           newHashTable[index + originalBucketNum] = new OrderedMap(highList);
         } else if (highList.length) {
           newHashTable[index + originalBucketNum] = new Vector(highList, false);
@@ -61,16 +65,12 @@ class HashMap<K, V> extends HashContainerBase<K> {
     }
     this.hashTable = newHashTable;
   }
-  clear() {
-    this.length = 0;
-    this.bucketNum = this.initBucketNum;
-    this.hashTable = [];
-  }
   forEach(callback: (element: [K, V], index: number) => void) {
+    const containers = Object.values(this.hashTable);
+    const containersNum = containers.length;
     let index = 0;
-    for (let i = 0; i < this.bucketNum; ++i) {
-      if (!this.hashTable[i]) continue;
-      this.hashTable[i].forEach(element => callback(element, index++));
+    for (let i = 0; i < containersNum; ++i) {
+      containers[i].forEach(element => callback(element, index++));
     }
   }
   /**
@@ -81,20 +81,21 @@ class HashMap<K, V> extends HashContainerBase<K> {
    */
   setElement(key: K, value: V) {
     const index = this.hashFunc(key) & (this.bucketNum - 1);
-    if (!this.hashTable[index]) {
+    const container = this.hashTable[index];
+    if (!container) {
       this.length += 1;
       this.hashTable[index] = new Vector([<[K, V]>[key, value]], false);
     } else {
-      const preSize = this.hashTable[index].size();
-      if (this.hashTable[index] instanceof Vector) {
-        for (const pair of this.hashTable[index]) {
+      const preSize = container.size();
+      if (container instanceof Vector) {
+        for (const pair of container) {
           if (pair[0] === key) {
             pair[1] = value;
             return;
           }
         }
-        (this.hashTable[index] as Vector<[K, V]>).pushBack([key, value]);
-        if (this.hashTable[index].size() >= HashMap.treeifyThreshold) {
+        (container as Vector<[K, V]>).pushBack([key, value]);
+        if (preSize + 1 >= HashMap.treeifyThreshold) {
           if (this.bucketNum <= HashMap.minTreeifySize) {
             this.length += 1;
             this.reAllocate();
@@ -102,9 +103,12 @@ class HashMap<K, V> extends HashContainerBase<K> {
           }
           this.hashTable[index] = new OrderedMap<K, V>(this.hashTable[index]);
         }
-      } else (this.hashTable[index] as OrderedMap<K, V>).setElement(key, value);
-      const curSize = this.hashTable[index].size();
-      this.length += curSize - preSize;
+        this.length += 1;
+      } else {
+        (container as OrderedMap<K, V>).setElement(key, value);
+        const curSize = container.size();
+        this.length += curSize - preSize;
+      }
     }
     if (this.length > this.bucketNum * HashMap.sigma) {
       this.reAllocate();
@@ -116,66 +120,63 @@ class HashMap<K, V> extends HashContainerBase<K> {
    */
   getElementByKey(key: K) {
     const index = this.hashFunc(key) & (this.bucketNum - 1);
-    if (!this.hashTable[index]) return undefined;
-    if (this.hashTable[index] instanceof OrderedMap) {
-      return (this.hashTable[index] as OrderedMap<K, V>).getElementByKey(key);
+    const container = this.hashTable[index];
+    if (!container) return undefined;
+    if (container instanceof OrderedMap) {
+      return (container as OrderedMap<K, V>).getElementByKey(key);
     } else {
-      for (const pair of this.hashTable[index]) {
+      for (const pair of container) {
         if (pair[0] === key) return pair[1];
       }
       return undefined;
     }
   }
-  /**
-   * @description Removes the element of the specified key.
-   * @param key The key you want to remove.
-   */
   eraseElementByKey(key: K) {
     const index = this.hashFunc(key) & (this.bucketNum - 1);
-    if (!this.hashTable[index]) return;
-    const preSize = this.hashTable[index].size();
-    if (this.hashTable[index] instanceof OrderedMap) {
-      (this.hashTable[index] as OrderedMap<K, V>).eraseElementByKey(key);
-      if (this.hashTable[index].size() <= HashMap.untreeifyThreshold) {
-        this.hashTable[index] = new Vector<[K, V]>(this.hashTable[index]);
-      }
-    } else {
+    const container = this.hashTable[index];
+    if (!container) return;
+    if (container instanceof Vector) {
       let pos = 0;
-      for (const pair of this.hashTable[index]) {
+      for (const pair of container) {
         if (pair[0] === key) {
-          this.hashTable[index].eraseElementByPos(pos);
-          break;
+          container.eraseElementByPos(pos);
+          this.length -= 1;
+          return;
         }
         pos += 1;
       }
+    } else {
+      const preSize = container.size();
+      (container as OrderedMap<K, V>).eraseElementByKey(key);
+      const curSize = container.size();
+      this.length += curSize - preSize;
+      if (curSize <= HashContainer.untreeifyThreshold) {
+        this.hashTable[index] = new Vector(container);
+      }
     }
-    const curSize = this.hashTable[index].size();
-    this.length += curSize - preSize;
   }
-  /**
-   * @param key The key you want to find.
-   * @return Boolean about if the specified element in the hash map.
-   */
   find(key: K) {
     const index = this.hashFunc(key) & (this.bucketNum - 1);
-    if (!this.hashTable[index]) return false;
-    if (this.hashTable[index] instanceof OrderedMap) {
-      return !(this.hashTable[index] as OrderedMap<K, V>).find(key)
-        .equals((this.hashTable[index] as OrderedMap<K, V>).end());
+    const container = this.hashTable[index];
+    if (!container) return false;
+    if (container instanceof OrderedMap) {
+      return !(container as OrderedMap<K, V>).find(key)
+        .equals((container as OrderedMap<K, V>).end());
     }
-    for (const pair of this.hashTable[index]) {
+    for (const pair of container) {
       if (pair[0] === key) return true;
     }
     return false;
   }
   [Symbol.iterator]() {
     return function * (this: HashMap<K, V>) {
-      let index = 0;
-      while (index < this.bucketNum) {
-        while (index < this.bucketNum && !this.hashTable[index]) index += 1;
-        if (index >= this.bucketNum) break;
-        for (const pair of this.hashTable[index]) yield pair;
-        index += 1;
+      const containers = Object.values(this.hashTable);
+      const containersNum = containers.length;
+      for (let i = 0; i < containersNum; ++i) {
+        const container = containers[i];
+        for (const element of container) {
+          yield element;
+        }
       }
     }.bind(this)();
   }